Fixes an issue with the fill-text-background mixin
There was an odd issue that was preventing the fill-text-background
mixin from working properly in a webkit based browser. If you flip
the statements in the supports condition everything seems to work
fine.

This also adjusts an issue where icons were getting cut off inside
of buttons due to a mismatch in line height.