Adds styles for go-table
This abstracts the styles out of the goponents table component and
brings them in to gosheets. This will help us begin to flush out
the new table design before we get to Angular proper. This only
includes the styles for the main table and not for any of the
controls so some alterations might still need to be made to make
this work consistently.